+++ tests/settings.test.ts
@@ -209,95 +209,3 @@ test.describe('locales', () => {
 		await expect(page.getByText('Servidor')).toBeVisible();
 	});
 });
-
-test('model select supports searching and filtering', async ({ page }) => {
-	await page.goto('/settings');
-	const modelComboBox = page.getByLabel('Available models');
-	await modelComboBox.click();
-
-	await expect(page.getByText('Recently used models', { exact: true })).toBeVisible();
-	await expect(page.getByText('Other models', { exact: true })).toBeVisible();
-
-	await modelComboBox.fill('gemma');
-	await expect(page.getByText('gemma:7b')).toBeVisible();
-	await expect(page.getByText('TheBloke/CodeLlama-70B-Python-AWQ')).not.toBeVisible();
-	await expect(page.getByText('openhermes2.5-mistral:latest')).not.toBeVisible();
-
-	await modelComboBox.fill('');
-	await expect(page.getByText('gemma:7b')).toBeVisible();
-	await expect(page.getByText('TheBloke/CodeLlama-70B-Python-AWQ')).toBeVisible();
-	await expect(page.getByText('openhermes2.5-mistral:latest')).toBeVisible();
-
-	await modelComboBox.fill('nonexistentmodel');
-	await expect(page.getByText('No results')).toBeVisible();
-});
-
-test('models are correctly grouped including recently used models', async ({ page }) => {
-	await page.goto('/settings');
-	await mockTagsResponse(page);
-	await chooseFromCombobox(page, 'Available models', MOCK_API_TAGS_RESPONSE.models[1].name);
-
-	// Mock a session with a specific model
-	await page.getByText('Sessions', { exact: true }).click();
-	await page.getByTestId('new-session').click();
-	await mockCompletionResponse(page, MOCK_SESSION_1_RESPONSE_1);
-	await page
-		.locator('.prompt-editor__textarea')
-		.fill('Who would win in a fight between Emma Watson and Jessica Alba?');
-	await page.getByText('Run').click();
-
-	await page.getByText('Settings', { exact: true }).click();
-
-	const modelComboBox = page.getByLabel('Available models');
-	await modelComboBox.click();
-
-	const recentlyUsedModelsGroup = page.locator('.field-combobox-group', {
-		hasText: 'Recently used models'
-	});
-	const recentlyUsedModel = recentlyUsedModelsGroup.locator('.field-combobox-item-label-option', {
-		hasText: MOCK_API_TAGS_RESPONSE.models[1].name
-	});
-	await expect(recentlyUsedModel).toBeVisible();
-
-	const otherModelsGroup = page.locator('.field-combobox-group', {
-		hasText: 'Other models'
-	});
-	await expect(otherModelsGroup).toContainText(MOCK_API_TAGS_RESPONSE.models[0].name);
-	await expect(otherModelsGroup).toContainText(MOCK_API_TAGS_RESPONSE.models[2].name);
-});
-
-test('models are correctly sorted by last used', async ({ page }) => {
-	await page.goto('/settings');
-	await mockTagsResponse(page);
-
-	await chooseFromCombobox(page, 'Available models', MOCK_API_TAGS_RESPONSE.models[1].name);
-
-	await page.getByText('Sessions', { exact: true }).click();
-	await page.getByTestId('new-session').click();
-	await mockCompletionResponse(page, MOCK_SESSION_1_RESPONSE_1);
-	await page
-		.locator('.prompt-editor__textarea')
-		.fill('Who would win in a fight between Emma Watson and Jessica Alba?');
-	await page.getByText('Run').click();
-
-	await page.getByTestId('new-session').click();
-	await chooseFromCombobox(page, 'Available models', MOCK_API_TAGS_RESPONSE.models[0].name);
-	await mockCompletionResponse(page, MOCK_SESSION_1_RESPONSE_1);
-	await page
-		.locator('.prompt-editor__textarea')
-		.fill('Who would win in a fight between Emma Watson and Jessica Alba?');
-	await page.getByText('Run').click();
-
-	await page.getByText('Settings', { exact: true }).click();
-
-	const modelComboBox = page.getByLabel('Available models');
-	await modelComboBox.click();
-
-	const recentlyUsedModelsGroup = page.locator('.field-combobox-group', {
-		hasText: 'Recently used models'
-	});
-	const recentlyUsedModels = recentlyUsedModelsGroup.locator('.field-combobox-item-label-option');
-	await expect(recentlyUsedModels).toHaveCount(2);
-	await expect(recentlyUsedModels.nth(0)).toHaveText(MOCK_API_TAGS_RESPONSE.models[0].name);
-	await expect(recentlyUsedModels.nth(1)).toHaveText(MOCK_API_TAGS_RESPONSE.models[1].name);
-});
