+++ tests/ui.test.ts
@@ -1,6 +1,12 @@
 import { expect, test } from '@playwright/test';
 
-import { MOCK_API_TAGS_RESPONSE, mockTagsResponse } from './utils';
+import {
+	chooseFromCombobox,
+	MOCK_API_TAGS_RESPONSE,
+	MOCK_SESSION_1_RESPONSE_1,
+	mockCompletionResponse,
+	mockTagsResponse
+} from './utils';
 
 test.describe('FieldSelect', () => {
 	test('filters options, shows selected value, and allows clearing', async ({ page }) => {
@@ -16,10 +22,17 @@ test.describe('FieldSelect', () => {
 		// Open the dropdown
 		await modelCombobox.click();
 
+		// Check that the groups are visible
+		await expect(page.getByText('Recently used models', { exact: true })).toBeVisible();
+		await expect(page.getByText('Other models', { exact: true })).toBeVisible();
+
 		// Check that all options are visible
 		await expect(page.getByRole('option')).toHaveCount(MOCK_API_TAGS_RESPONSE.models.length);
 
 		// Filter options
+		await modelCombobox.fill('nonexistentmodel');
+		await expect(page.getByText('No results')).toBeVisible();
+
 		await modelCombobox.fill('llama');
 		await expect(page.getByRole('option')).toHaveCount(1);
 		await expect(page.getByTitle('Clear')).not.toBeVisible();
@@ -65,5 +78,52 @@ test.describe('FieldSelect', () => {
 		await expect(modelCombobox).toHaveAttribute('placeholder', 'Search');
 		await expect(page.getByRole('option')).toHaveCount(0);
 		await expect(page.getByTitle('Clear')).not.toBeVisible();
+	});
+
+	test('models are correctly grouped and sorted', async ({ page }) => {
+		const newSessionButton = page.getByTestId('new-session');
+		await page.goto('/settings');
+		await mockTagsResponse(page);
+
+		await page.getByText('Sessions', { exact: true }).click();
+		await newSessionButton.click();
+
+		// Select a model
+		await chooseFromCombobox(page, 'Available models', MOCK_API_TAGS_RESPONSE.models[1].name);
+
+		// Mock a session with a specific model
+		await mockCompletionResponse(page, MOCK_SESSION_1_RESPONSE_1);
+		await page
+			.locator('.prompt-editor__textarea')
+			.fill('Who would win in a fight between Emma Watson and Jessica Alba?');
+		await page.getByText('Run').click();
+
+		// Create another session with a different model
+		await newSessionButton.click();
+		await chooseFromCombobox(page, 'Available models', MOCK_API_TAGS_RESPONSE.models[0].name);
+		await mockCompletionResponse(page, MOCK_SESSION_1_RESPONSE_1);
+		await page
+			.locator('.prompt-editor__textarea')
+			.fill('Who would win in a fight between Emma Watson and Jessica Alba?');
+		await page.getByText('Run').click();
+
+		await newSessionButton.click();
+		const modelCombobox = page.getByLabel('Available models');
+		await modelCombobox.click();
+
+		// Check for "Recently used models" group
+		const recentlyUsedModelsGroup = page.locator('.field-combobox-group', {
+			hasText: 'Recently used models'
+		});
+		const recentlyUsedModels = recentlyUsedModelsGroup.locator('.field-combobox-item-label-option');
+		await expect(recentlyUsedModels).toHaveCount(2);
+		await expect(recentlyUsedModels.nth(0)).toHaveText(MOCK_API_TAGS_RESPONSE.models[0].name);
+		await expect(recentlyUsedModels.nth(1)).toHaveText(MOCK_API_TAGS_RESPONSE.models[1].name);
+
+		// Check for "Other models" group
+		const otherModelsGroup = page.locator('.field-combobox-group', {
+			hasText: 'Other models'
+		});
+		await expect(otherModelsGroup).toContainText(MOCK_API_TAGS_RESPONSE.models[2].name);
 	});
 });
